requirements.txt
langchain
pypdf
chromadb
openai
tiktoken
python-dotenv
lark
unstructured
tabulate
pdf2image
pytesseract
